+++ demisto_sdk/commands/common/hook_validations/integration.py
@@ -18,6 +18,8 @@ from demisto_sdk.commands.common.errors import (FOUND_FILES_AND_ERRORS,
                                                 FOUND_FILES_AND_IGNORED_ERRORS,
                                                 Errors)
 from demisto_sdk.commands.common.handlers import JSON_Handler, YAML_Handler
+from demisto_sdk.commands.common.hook_validations.base_validator import \
+    error_codes
 from demisto_sdk.commands.common.hook_validations.content_entity_validator import \
     ContentEntityValidator
 from demisto_sdk.commands.common.hook_validations.description import \
@@ -50,6 +52,7 @@ class IntegrationValidator(ContentEntityValidator):
                          json_file_path=json_file_path, skip_docker_check=skip_docker_check)
         self.validate_all = validate_all
 
+    @error_codes('BA100')
     def is_valid_version(self):
         # type: () -> bool
         if self.current_file.get("commonfields", {}).get('version') == self.DEFAULT_VERSION:
@@ -70,17 +73,17 @@ class IntegrationValidator(ContentEntityValidator):
             return True
 
         answers = [
-            self.is_changed_context_path(),
-            self.is_removed_integration_parameters(),
-            self.is_added_required_fields(),
-            self.is_changed_command_name_or_arg(),
-            self.is_changed_subtype(),
-            self.is_not_valid_display_configuration(),
-            self.is_changed_removed_yml_fields(),
+            self.no_change_to_context_path(),
+            self.no_removed_integration_parameters(),
+            self.no_added_required_fields(),
+            self.no_changed_command_name_or_arg(),
+            self.no_changed_subtype(),
+            self.is_valid_display_configuration(),
+            self.no_changed_removed_yml_fields(),
             # will move to is_valid_integration after https://github.com/demisto/etc/issues/17949
-            not self.is_outputs_for_reputations_commands_valid(),
+            self.is_outputs_for_reputations_commands_valid(),
         ]
-        return not any(answers)
+        return all(answers)
 
     def core_integration_validations(self, validate_rn: bool = True):
         """Perform the core integration validations (common to both beta and regular integrations)
@@ -184,6 +187,7 @@ class IntegrationValidator(ContentEntityValidator):
         ]
         return all(answers)
 
+    @error_codes('IN140')
     def is_unskipped_integration(self, conf_json_data):
         """Validated the integration testing is not skipped."""
         skipped_integrations = conf_json_data.get('skipped_integrations', {})
@@ -195,6 +199,7 @@ class IntegrationValidator(ContentEntityValidator):
                 self.is_valid = False
         return self.is_valid
 
+    @error_codes('IN127')
     def _is_valid_deprecated_integration_display_name(self) -> bool:
         is_valid = True
         is_deprecated = self.current_file.get('deprecated', False)
@@ -206,6 +211,7 @@ class IntegrationValidator(ContentEntityValidator):
                     is_valid = False
         return is_valid
 
+    @error_codes('IN128')
     def _is_valid_deprecated_integration_description(self) -> bool:
         is_valid = True
         is_deprecated = self.current_file.get('deprecated', False)
@@ -230,6 +236,7 @@ class IntegrationValidator(ContentEntityValidator):
         tests = self.current_file.get('tests', [])
         return self.are_tests_registered_in_conf_json_file_or_yml_file(tests)
 
+    @error_codes('IN100,IN101,IN102,IN103')
     def is_valid_param(self, param_name, param_display):
         # type: (str, str) -> bool
         """Check if the given parameter has the right configuration."""
@@ -307,6 +314,7 @@ class IntegrationValidator(ContentEntityValidator):
             return False
         return True
 
+    @error_codes('IN102')
     def is_valid_checkbox_param(self, configuration_param, param_name):
         # type: (dict, str) -> bool
         """Check if the given checkbox parameter required field is False.
@@ -319,6 +327,7 @@ class IntegrationValidator(ContentEntityValidator):
                 return False
         return True
 
+    @error_codes('IN104')
     def is_valid_category(self):
         # type: () -> bool
         """Check that the integration category is in the schema."""
@@ -331,6 +340,7 @@ class IntegrationValidator(ContentEntityValidator):
 
         return True
 
+    @error_codes('IN105,IN144,IN106')
     def is_valid_default_array_argument_in_reputation_command(self):
         # type: () -> bool
         """Check if a reputation command (domain/email/file/ip/url/cve)
@@ -375,6 +385,7 @@ class IntegrationValidator(ContentEntityValidator):
             print_error(Errors.suggest_fix(self.file_path))
         return flag
 
+    @error_codes('IN134')
     def is_valid_default_argument(self):
         # type: () -> bool
         """Check if a  command has at most 1 default argument.
@@ -420,6 +431,7 @@ class IntegrationValidator(ContentEntityValidator):
 
         return missing_outputs, missing_descriptions
 
+    @error_codes('DB100,DB101,IN107')
     def is_outputs_for_reputations_commands_valid(self):
         # type: () -> bool
         """Check if a reputation command (domain/email/file/ip/url)
@@ -458,7 +470,9 @@ class IntegrationValidator(ContentEntityValidator):
                         error_message, error_code = Errors.dbot_invalid_description(command_name,
                                                                                     missing_descriptions,
                                                                                     context_standard)
-                        self.handle_error(error_message, error_code, file_path=self.file_path, warning=True)
+                        if self.handle_error(error_message, error_code, file_path=self.file_path, warning=True):
+                            self.is_valid = False
+                            output_for_reputation_valid = False
 
                 # validate the IOC output
                 reputation_output = IOC_OUTPUTS_DICT.get(command_name)
@@ -472,6 +486,7 @@ class IntegrationValidator(ContentEntityValidator):
 
         return output_for_reputation_valid
 
+    @error_codes('IN108')
     def is_valid_subtype(self):
         # type: () -> bool
         """Validate that the subtype is python2 or python3."""
@@ -486,9 +501,11 @@ class IntegrationValidator(ContentEntityValidator):
 
         return True
 
-    def is_changed_subtype(self):
+    @error_codes('BC100')
+    def no_changed_subtype(self):
         # type: () -> bool
-        """Validate that the subtype was not changed."""
+        """Validate that the subtype was not changed.
+        Returns True if valid, and False otherwise."""
         type_ = self.current_file.get('script', {}).get('type')
         if type_ == 'python':
             subtype = self.current_file.get('script', {}).get('subtype')
@@ -499,9 +516,9 @@ class IntegrationValidator(ContentEntityValidator):
                     if self.handle_error(error_message, error_code, file_path=self.file_path,
                                          warning=self.structure_validator.quiet_bc):
                         self.is_valid = False
-                        return True
+                        return False
 
-        return False
+        return True
 
     def is_valid_beta(self):
         # type: () -> bool
@@ -516,6 +533,7 @@ class IntegrationValidator(ContentEntityValidator):
                 valid_status = False
         return valid_status
 
+    @error_codes('IN109')
     def _id_has_no_beta_substring(self):
         # type: () -> bool
         """Checks that 'id' field dose not include the substring 'beta'"""
@@ -528,6 +546,7 @@ class IntegrationValidator(ContentEntityValidator):
 
         return True
 
+    @error_codes('IN110')
     def _name_has_no_beta_substring(self):
         # type: () -> bool
         """Checks that 'name' field dose not include the substring 'beta'"""
@@ -539,6 +558,7 @@ class IntegrationValidator(ContentEntityValidator):
 
         return True
 
+    @error_codes('IN111')
     def _has_beta_param(self):
         # type: () -> bool
         """Checks that integration has 'beta' field with value set to true"""
@@ -550,6 +570,7 @@ class IntegrationValidator(ContentEntityValidator):
 
         return True
 
+    @error_codes('IN112')
     def _is_display_contains_beta(self):
         # type: () -> bool
         """Checks that 'display' field includes the substring 'beta'"""
@@ -562,6 +583,7 @@ class IntegrationValidator(ContentEntityValidator):
 
         return True
 
+    @error_codes('IN113')
     def has_no_duplicate_args(self):
         # type: () -> bool
         """Check if a command has the same arg more than once
@@ -590,6 +612,7 @@ class IntegrationValidator(ContentEntityValidator):
 
         return does_not_have_duplicate_args
 
+    @error_codes('IN139')
     def no_incident_in_core_packs(self):
         """check if commands' name or argument contains the word incident"""
 
@@ -617,6 +640,7 @@ class IntegrationValidator(ContentEntityValidator):
 
         return no_incidents
 
+    @error_codes('IN114')
     def has_no_duplicate_params(self):
         # type: () -> bool
         """Check if the integration has the same param more than once
@@ -659,12 +683,13 @@ class IntegrationValidator(ContentEntityValidator):
                 command_to_args[command['name']][arg['name']] = arg.get('required', False)
         return command_to_args
 
-    def is_changed_command_name_or_arg(self):
+    @error_codes('BC104')
+    def no_changed_command_name_or_arg(self):
         # type: () -> bool
         """Check if a command name or argument as been changed.
 
         Returns:
-            bool. Whether a command name or argument as been changed.
+            bool. True if valid, and False otherwise.
         """
         changed_commands = []
         current_command_to_args = self._get_command_to_args(self.current_file)
@@ -679,8 +704,8 @@ class IntegrationValidator(ContentEntityValidator):
             if self.handle_error(error_message, error_code, file_path=self.file_path,
                                  warning=self.structure_validator.quiet_bc):
                 self.is_valid = False
-                return True
-        return False
+                return False
+        return True
 
     @staticmethod
     def _is_sub_set(supposed_bigger_list, supposed_smaller_list):
@@ -688,6 +713,7 @@ class IntegrationValidator(ContentEntityValidator):
         """Check if supposed_smaller_list is a subset of the supposed_bigger_list"""
         return all(item in supposed_bigger_list for item in supposed_smaller_list)
 
+    @error_codes('IN115')
     def _get_command_to_context_paths(self, integration_json):
         # type: (dict) -> dict
         """Get a dictionary command name to it's context paths.
@@ -717,22 +743,23 @@ class IntegrationValidator(ContentEntityValidator):
             command_to_context_dict[command['name']] = sorted(context_list)
         return command_to_context_dict
 
-    def is_changed_context_path(self):
+    @error_codes('BC102')
+    def no_change_to_context_path(self):
         # type: () -> bool
         """Check if a context path as been changed.
 
         Returns:
-            bool. Whether a context path as been changed.
+            bool. True if valid, and False otherwise.
         """
         current_command_to_context_paths = self._get_command_to_context_paths(self.current_file)
         old_command_to_context_paths = self._get_command_to_context_paths(self.old_file)
         # if old integration command has no outputs, no change of context will occur.
         if not old_command_to_context_paths:
-            return False
+            return True
         # if new integration command has no outputs, and old one does, a change of context will occur.
         if not current_command_to_context_paths and old_command_to_context_paths \
                 and not self.structure_validator.quiet_bc:
-            return True
+            return False
         for old_command, old_context_paths in old_command_to_context_paths.items():
             if old_command in current_command_to_context_paths.keys():
                 if not self._is_sub_set(current_command_to_context_paths[old_command], old_context_paths):
@@ -740,14 +767,17 @@ class IntegrationValidator(ContentEntityValidator):
                     if self.handle_error(error_message, error_code, file_path=self.file_path,
                                          warning=self.structure_validator.quiet_bc):
                         self.is_valid = False
-                        return True
+                        return False
 
-        return False
+        return True
 
-    def is_removed_integration_parameters(self):
+    @error_codes('IN129')
+    def no_removed_integration_parameters(self):
         # type: () -> bool
-        """Check if integration parameters were removed."""
-        is_removed_parameter = False
+        """Check if integration parameters were removed.
+        Returns True if valid, and False otherwise.
+        """
+        no_removed_parameter = True
         current_configuration = self.current_file.get('configuration', [])
         old_configuration = self.old_file.get('configuration', [])
         current_param_names = {param.get('name') for param in current_configuration}
@@ -758,9 +788,9 @@ class IntegrationValidator(ContentEntityValidator):
             if self.handle_error(error_message, error_code, file_path=self.file_path,
                                  warning=self.structure_validator.quiet_bc):
                 self.is_valid = False
-                is_removed_parameter = True
+                no_removed_parameter = False
 
-        return is_removed_parameter
+        return no_removed_parameter
 
     @staticmethod
     def _get_field_to_required_dict(integration_json):
@@ -778,8 +808,11 @@ class IntegrationValidator(ContentEntityValidator):
             field_to_required[field.get('name')] = field.get('required', False)
         return field_to_required
 
-    def is_changed_removed_yml_fields(self):
-        """checks if some specific Fields in the yml file were changed from true to false or removed"""
+    @error_codes('IN147')
+    def no_changed_removed_yml_fields(self):
+        """checks if some specific Fields in the yml file were changed from true to false or removed
+        Returns True if valid, and False otherwise.
+        """
         fields = ['feed', 'isfetch', 'longRunning', 'longRunningPort', 'ismappable', 'isremotesyncin',
                   'isremotesyncout']
         currentscript = self.current_file.get('script', {})
@@ -802,15 +835,18 @@ class IntegrationValidator(ContentEntityValidator):
             if self.handle_error(error_message, error_code, file_path=self.file_path,
                                  warning=self.structure_validator.quiet_bc):
                 self.is_valid = False
-                return True
-        return False
+                return False
+        return True
 
-    def is_added_required_fields(self):
+    @error_codes('IN116')
+    def no_added_required_fields(self):
         # type: () -> bool
-        """Check if required field were added."""
+        """Check if required field were added.
+        Returns True if valid, and False otherwise.
+        """
         current_field_to_required = self._get_field_to_required_dict(self.current_file)
         old_field_to_required = self._get_field_to_required_dict(self.old_file)
-        is_added_required = False
+        no_added_required = True
         for field, required in current_field_to_required.items():
             if field in old_field_to_required.keys():
                 # if required is True and old_field is False.
@@ -819,7 +855,7 @@ class IntegrationValidator(ContentEntityValidator):
                     if self.handle_error(error_message, error_code, file_path=self.file_path,
                                          warning=self.structure_validator.quiet_bc):
                         self.is_valid = False
-                        is_added_required = True
+                        no_added_required = False
 
             # if required is True but no old field.
             elif required:
@@ -827,23 +863,25 @@ class IntegrationValidator(ContentEntityValidator):
                 if self.handle_error(error_message, error_code, file_path=self.file_path,
                                      warning=self.structure_validator.quiet_bc):
                     self.is_valid = False
-                    is_added_required = True
+                    no_added_required = False
 
-        return is_added_required
+        return no_added_required
 
     def is_id_equals_name(self):
         """Check whether the integration's ID is equal to its name
 
         Returns:
-            bool. Whether the integration's id equals to its name
+            bool. True if valid, and False otherwise.
         """
         return super(IntegrationValidator, self)._is_id_equals_name('integration')
 
-    def is_not_valid_display_configuration(self):
+    @error_codes('IN117,IN118')
+    def is_valid_display_configuration(self):
         """Validate that the display settings are not empty for non-hidden fields and for type 17 params.
 
         Returns:
             bool. Whether the display is there for non-hidden fields.
+            Returns True if valid, and False otherwise.
         """
         configuration = self.current_file.get('configuration', [])
         for configuration_param in configuration:
@@ -858,7 +896,7 @@ class IntegrationValidator(ContentEntityValidator):
                     if self.handle_error(error_message, error_code, file_path=self.file_path,
                                          warning=self.structure_validator.quiet_bc):
                         self.is_valid = False
-                        return True
+                        return False
 
             elif not is_field_hidden and not configuration_display and not configuration_param.get('displaypassword') \
                     and configuration_param['name'] not in ('feedExpirationPolicy', 'feedExpirationInterval'):
@@ -866,9 +904,9 @@ class IntegrationValidator(ContentEntityValidator):
                 if self.handle_error(error_message, error_code, file_path=self.file_path,
                                      warning=self.structure_validator.quiet_bc):
                     self.is_valid = False
-                    return True
+                    return False
 
-        return False
+        return True
 
     def is_docker_image_valid(self):
         # type: () -> bool
@@ -881,7 +919,8 @@ class IntegrationValidator(ContentEntityValidator):
                                                       print_as_warnings=self.print_as_warnings,
                                                       suppress_print=self.suppress_print,
                                                       json_file_path=self.json_file_path,
-                                                      is_iron_bank=is_iron_bank)
+                                                      is_iron_bank=is_iron_bank,
+                                                      specific_validations=self.specific_validations)
 
         # making sure we don't show error of validation if fetching is failed.
         _, error_code = Errors.docker_tag_not_fetched('', '')
@@ -894,6 +933,7 @@ class IntegrationValidator(ContentEntityValidator):
         self.is_valid = False
         return False
 
+    @error_codes('IN119')
     def is_valid_feed(self):
         # type: () -> bool
         valid_from_version = valid_feed_params = True
@@ -908,6 +948,7 @@ class IntegrationValidator(ContentEntityValidator):
             valid_feed_params = self.all_feed_params_exist()
         return valid_from_version and valid_feed_params
 
+    @error_codes('IN120')
     def is_valid_pwsh(self) -> bool:
         if self.current_file.get('script', {}).get('type') == TYPE_PWSH:
             from_version = self.current_file.get('fromversion', DEFAULT_CONTENT_ITEM_FROM_VERSION)
@@ -918,6 +959,7 @@ class IntegrationValidator(ContentEntityValidator):
                     return False
         return True
 
+    @error_codes('IN121,IN148')
     def is_valid_fetch(self) -> bool:
         """
         validate that all required fields in integration that have fetch incidents are in the yml file.
@@ -955,6 +997,7 @@ class IntegrationValidator(ContentEntityValidator):
 
         return fetch_params_exist
 
+    @error_codes('IN126,IN125')
     def is_valid_max_fetch_and_first_fetch(self) -> bool:
         """
         validate that the max_fetch and first_fetch params exist in the yml and the max_fetch has default value
@@ -992,6 +1035,7 @@ class IntegrationValidator(ContentEntityValidator):
 
         return fetch_params_exist
 
+    @error_codes('IN122')
     def all_feed_params_exist(self) -> bool:
         """
         validate that all required fields in feed integration are in the yml file.
@@ -1030,6 +1074,7 @@ class IntegrationValidator(ContentEntityValidator):
 
         return params_exist
 
+    @error_codes('IN123')
     def is_valid_display_name(self):
         # type: () -> bool
         version_number: Optional[str] = get_file_version_suffix_if_exists(self.current_file,
@@ -1046,6 +1091,7 @@ class IntegrationValidator(ContentEntityValidator):
 
             return True
 
+    @error_codes('IN150')
     def is_valid_display_name_for_siem(self) -> bool:
         is_siem = self.current_file.get('script', {}).get('isFetchEvents')
 
@@ -1058,6 +1104,7 @@ class IntegrationValidator(ContentEntityValidator):
 
         return True
 
+    @error_codes('IN124')
     def is_valid_hidden_params(self) -> bool:
         """
         Verify there are no non-allowed hidden integration parameters.
@@ -1084,7 +1131,8 @@ class IntegrationValidator(ContentEntityValidator):
         """
         image_validator = ImageValidator(self.file_path, ignored_errors=self.ignored_errors,
                                          print_as_warnings=self.print_as_warnings,
-                                         json_file_path=self.json_file_path)
+                                         json_file_path=self.json_file_path,
+                                         specific_validations=self.specific_validations)
         if not image_validator.is_valid():
             return False
         return True
@@ -1097,7 +1145,8 @@ class IntegrationValidator(ContentEntityValidator):
         """
         description_validator = DescriptionValidator(self.file_path, ignored_errors=self.ignored_errors,
                                                      print_as_warnings=self.print_as_warnings,
-                                                     json_file_path=self.json_file_path)
+                                                     json_file_path=self.json_file_path,
+                                                     specific_validations=self.specific_validations)
         if beta_integration:
             if not description_validator.is_valid_beta_description():
                 return False
@@ -1106,6 +1155,7 @@ class IntegrationValidator(ContentEntityValidator):
                 return False
         return True
 
+    @error_codes('IN130')
     def is_there_a_runnable(self) -> bool:
         """Verifies there's at least one runnable command.
             at least one of:
@@ -1129,6 +1179,7 @@ class IntegrationValidator(ContentEntityValidator):
             return False
         return True
 
+    @error_codes('IN135')
     def is_valid_parameters_display_name(self) -> bool:
         """Verifies integration parameters display name is valid.
 
@@ -1151,6 +1202,7 @@ class IntegrationValidator(ContentEntityValidator):
 
         return True
 
+    @error_codes('IN138,IN137')
     def is_valid_integration_file_path(self) -> bool:
         absolute_file_path = self.file_path
         integrations_folder = os.path.basename(os.path.dirname(absolute_file_path))
@@ -1177,6 +1229,7 @@ class IntegrationValidator(ContentEntityValidator):
 
         return True
 
+    @error_codes('IN131')
     def is_mapping_fields_command_exist(self) -> bool:
         """
         Check if get-mapping-fields command exists in the YML if  the ismappble field is set to true
@@ -1193,6 +1246,7 @@ class IntegrationValidator(ContentEntityValidator):
                     return False
         return True
 
+    @error_codes('RM102,IN136')
     def is_context_correct_in_readme(self) -> bool:
         """
         Checks if there has been a corresponding change to the integration's README
@@ -1261,6 +1315,7 @@ class IntegrationValidator(ContentEntityValidator):
 
         return all(answers)
 
+    @error_codes('BA108')
     def check_separators_in_folder(self) -> bool:
         """
         Check if there are separators in the integration folder.
@@ -1281,6 +1336,7 @@ class IntegrationValidator(ContentEntityValidator):
 
         return True
 
+    @error_codes('BA109')
     def check_separators_in_files(self):
         """
         Check if there are separators in the integration files names.
@@ -1322,6 +1378,7 @@ class IntegrationValidator(ContentEntityValidator):
 
         return True
 
+    @error_codes('BA110')
     def name_not_contain_the_type(self):
         """
         Check that the entity name or display name does not contain the entity type
@@ -1363,6 +1420,7 @@ class IntegrationValidator(ContentEntityValidator):
         endpoint_command = [arg for arg in commands if arg.get('name') == 'endpoint'][0]
         return self._is_valid_endpoint_inputs(endpoint_command, required_arguments=ENDPOINT_FLEXIBLE_REQUIRED_ARGS)
 
+    @error_codes('IN141,IN105')
     def _is_valid_endpoint_inputs(self, command_data, required_arguments):
         """
         Check if the input for endpoint commands includes at least one required_arguments,
@@ -1395,6 +1453,7 @@ class IntegrationValidator(ContentEntityValidator):
                 return False
         return True
 
+    @error_codes('IN142,IN143')
     def default_params_have_default_additional_info(self):
         """Check if the all integration params that can have a default description have a it set.
         Raises warnings if the additional info is defined (not empty) but is different from the default.
@@ -1427,6 +1486,7 @@ class IntegrationValidator(ContentEntityValidator):
             return False
         return True
 
+    @error_codes('IN146')
     def has_no_fromlicense_key_in_contributions_integration(self):
         """Verifies that only xsoar supported integration can contain the `fromlicense` key in the configuration.
 
@@ -1456,6 +1516,7 @@ class IntegrationValidator(ContentEntityValidator):
             raise Exception('Could not find the pack name of the integration, '
                             'please verify the integration is in a pack')
 
+    @error_codes('IN145')
     def is_api_token_in_credential_type(self):
         """Checks if there are no keys with the `encrypted` type,
             The best practice is to use the `credential` type instead of `encrypted`.
@@ -1483,6 +1544,7 @@ class IntegrationValidator(ContentEntityValidator):
         raise Exception('Could not find the pack name of the integration, '
                         'please verify the integration is in a pack')
 
+    @error_codes('IN149')
     def are_common_outputs_with_description(self):
         defaults = json.loads(
             (Path(__file__).absolute().parents[2] / 'common/default_output_descriptions.json').read_text())
@@ -1504,6 +1566,7 @@ class IntegrationValidator(ContentEntityValidator):
 
         return True
 
+    @error_codes('RM110')
     def verify_yml_commands_match_readme(self, is_modified=False):
         """
         Checks if there are commands that doesn't appear in the readme but appear in the .yml file